app/requirements.txt
tensorflow-cpu
numpy == 1.18.0
streamlit
opencv-contrib-python-headless
Pillow
keras